+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- CHANGE: Rename `Chmod` to `Metadata`. [#179], [#180], previously [#112], [#161]
 - META: Rename `v4-legacy` branch to `main`, to further clarify status and potentially prepare for a breaking release.
 - DOCS: Change `v5` to `vNext` to allow for a breaking release.